Fix build

+++ inference-chain/x/inference/module/top_miners.go
@@ -54,7 +54,7 @@ func (am AppModule) RegisterTopMiners(ctx context.Context, participants []*types
 				return err
 			}
 			topMinerVestingPeriod := &params.TokenomicsParams.TopMinerVestingPeriod
-			err = am.keeper.PayParticipantFromModule(ctx, typedAction.Miner.Address, uint64(typedAction.Payout), types.TopRewardPoolAccName, "top_miner", topMinerVestingPeriod)
+			err = am.keeper.PayParticipantFromModule(ctx, typedAction.Miner.Address, typedAction.Payout, types.TopRewardPoolAccName, "top_miner", topMinerVestingPeriod)
 			if err != nil {
 				return err
 			}
